fix(route): 优惠券、sku修改
(1)重构sku重复代码
(2)修复寻找最优优惠券的逻辑
(3)修复route未过滤删除标志
